feat: CLIN-910 revamp clinical signs wip
