+++ src/components/Prescription/StepsPanel/index.tsx
@@ -1,13 +1,16 @@
 import { Space, Steps, Typography } from 'antd';
 
+import { useAppDispatch } from 'store';
 import { useGlobals } from 'store/global';
 import { usePrescriptionForm } from 'store/prescription';
+import { prescriptionFormActions } from 'store/prescription/slice';
 
 import styles from './index.module.scss';
 
 const { Title } = Typography;
 
 const StepsPanel = () => {
+  const dispatch = useAppDispatch();
   const { getAnalysisNameByCode } = useGlobals();
   const { config, currentStep } = usePrescriptionForm();
 
@@ -18,16 +21,19 @@ const StepsPanel = () => {
         <Title level={4}>{getAnalysisNameByCode(config?.analysisTitle!, false)}</Title>
       </Space>
       <Steps direction="vertical" size="small" current={currentStep?.index}>
-        {/**
-          onStepClick={(index) => {
+        {config?.steps.map((step) => (
+          <Steps.Step
+            onStepClick={(index) => {
               dispatch(
                 prescriptionFormActions.goTo({
                   index,
                 }),
               );
-            }} */}
-        {config?.steps.map((step) => (
-          <Steps.Step className={styles.stepsItem} key={step.index} title={step.title} />
+            }}
+            className={styles.stepsItem}
+            key={step.index}
+            title={step.title}
+          />
         ))}
       </Steps>
     </Space>
