+++ src/components/Prescription/components/ClinicalSignsSelect/NotObservedSignsList.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const ObservedSignsList = () => <></>;
+
+export default ObservedSignsList;
