+++ src/components/Prescription/components/ClinicalSignsSelect/index.module.scss
@@ -17,7 +17,7 @@ $label-width: 350px;
   }
 
   .hpoFormItem {
-    margin-bottom: 12px;
+    margin-bottom: 4px;
     :global(.ant-form-item) {
       margin-bottom: 0;
     }
